requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
asttokens==3.0.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap5==2026.3
cron-descriptor==1.4.5
cryptography==46.0.5
cssselect2==0.9.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.2
django-allauth==65.14.3
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-crispy-forms==2.6
django-htmx==1.27.0
django-idempotency-key==1.3.0
django-phonenumber-field==8.4.0
django-redis==6.0.0
django-timezone-field==7.2.1
django-tinymce==5.0.0
django_celery_results==2.6.0
djangorestframework==3.16.1
docopt==0.6.2
drf-spectacular==0.29.0
exa-py==1.16.1
executing==2.2.1
Faker==40.5.1
filetype==1.2.0
fonttools==4.61.1
freetype-py==2.5.1
frozenlist==1.8.0
google-api-core==2.8.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-genai==1.65.0
googleapis-common-protos==1.56.1
greenlet==3.3.2
groq==0.37.1
gunicorn==25.1.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
inflection==0.5.1
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwt==1.4.0
kiwisolver==1.4.9
kombu==5.6.2
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.16
langchain-deepseek==1.0.1
langchain-exa==1.0.0
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-openai==1.1.10
langchain-tavily==0.2.17
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.9
lxml==6.0.2
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.2
oauthlib==3.3.1
openai==2.24.0
orjson==3.11.7
ormsgpack==1.12.2
oscrypto==1.3.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pdfminer.six==20251230
pdfplumber==0.11.9
phonenumbers==9.0.25
pillow==12.1.1
pipreqs==0.4.13
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf>=4.21.0,<5.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
pyotp==2.9.0
pyparsing==3.3.2
pypdf==6.7.4
pypdfium2==5.5.0
python-bidi==0.6.7
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
qrcode==8.2
redis==7.2.1
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rlPyCairo==0.4.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.47
sqlite-vec==0.1.6
sqlparse==0.5.5
stack-data==0.6.3
svglib==1.6.0
tavily==1.1.0
tenacity==9.1.4
tiktoken==0.12.0
tinycss2==1.5.1
tqdm==4.67.3
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.3
uuid_utils==0.14.1
vine==5.1.0
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
whitenoise==6.12.0
xhtml2pdf==0.2.17
xxhash==3.6.0
yarg==0.1.10
yarl==1.22.0
zstandard==0.25.0
